backend/requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
pillow==11.1.0
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
shortuuid==1.0.13
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
gunicorn==20.1.0
pytz==2022.7.1
whitenoise==6.4.0
yapf==0.32.0
